Client: Update caching of conversation user data to be stored in message portion of redux state; update logic to determine whether to fetch or retrieved cached user data in conversation component
