+++ client/src/redux/user/user.reducer.ts
@@ -1,4 +1,5 @@
 import { Map } from 'immutable';
+import { UserInfoMap } from '../../components/conversation/conversation.component';
 import { UserActions, UserActionTypes, UserState } from './user.types';
 import { addUserToUsersArray } from './user.utils';
 
@@ -26,7 +27,6 @@ const INITIAL_STATE: UserState = {
   deleteAccountError: null,
   feedPostReactingUsers: null,
   conversationUsers: null,
-  conversationToUserDataMap: Map([]),
 };
 
 const userReducer = (
@@ -92,14 +92,6 @@ const userReducer = (
           action.payload
         ),
         otherUserError: null,
-      };
-    case UserActions.ADD_TO_CONVERSATION_TO_USER_DATA_MAP:
-      return {
-        ...state,
-        conversationToUserDataMap: state.conversationToUserDataMap.set(
-          action.payload.conversationId,
-          action.payload.userData
-        ),
       };
     case UserActions.SIGN_UP_SUCCESS:
       return {
